# #8594 Optimize Project Euler solutions