## Describe your change:

### Optimized these solutions 

* [ ] Problem 010 / sol2.py (0.48s)
* [ ] Problem 010 / sol1.py (0.44s)
* [ ] Problem 135 / sol1.py (0.14s)
* [ ] Problem 073 / sol1.py (0.14s)
* [ ] Problem 034 / sol1.py (0.22s)

### These solutions are well optimized 

* [ ] Problem 078 / sol1.py (3.67s)
* [ ] Problem 145 / sol1.py (25.00s)
* [ ] Problem 070 / sol1.py (13.86s)
* [ ] Problem 104 / sol1.py (6.84s)
* [ ] Problem 092 / sol1.py (6.01s)


### Checklist:
* [ ] I have read [CONTRIBUTING.md](https://github.com/TheAlgorithms/Python/blob/master/CONTRIBUTING.md).
* [ ] This pull request is all my own work -- I have not plagiarized.
* [ ] I know that pull requests will not be merged if they fail the automated tests.
* [ ] This PR only changes one algorithm file.  To ease review, please open separate PRs for separate algorithms.
* [ ] All new Python files are placed inside an existing directory.
* [ ] All filenames are in all lowercase characters with no spaces or dashes.
* [ ] All functions and variable names follow Python naming conventions.
* [ ] All function parameters and return values are annotated with Python [type hints](https://docs.python.org/3/library/typing.html).
* [ ] All functions have [doctests](https://docs.python.org/3/library/doctest.html) that pass the automated testing.
* [ ] All new algorithms include at least one URL that points to Wikipedia or another similar explanation.
* [ ] If this pull request resolves one or more open issues then the commit message contains `Fixes: #{$ISSUE_NO}`.